backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = ""
authors = [
    {name = "bill.davis",email = "1972806+peoplesmeat@users.noreply.github.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard] (>=0.44.0,<0.45.0)",
    "pymssql>=2.3.1",
    "sqlalchemy>=2.0.0",
    "python-dotenv>=1.0.0",
]

[tool.poetry]
packages = [{include = "backend", from = "src"}]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
